requirements.txt
customtkinter
sqlalchemy
scikit-learn
numpy
pyTelegramBotAPI
python-dotenv
langchain
langchain-community
langchain-core
langchain-ollama
langchain-chroma
langchain-text-splitters
chromadb
opencv-python
ultralytics
reportlab
pypdf
pillow
requests
pyyaml
roboflow
